Basic example of the new group features.
Version 2.002.
Containes the new group movement and group structure features.
Is intended to make things easier, as sending commands to groups and
making
groups. The structures are mentioned to ensure their own consistency.
Group formations, A* and opponent modelling are not included yet.
The code doesnt represent the actual behavior of the final AI. It is
mentioned to demonstrate the new functionality. Feel free to experiment
and add features.